fix to choose JS redirection explicitly
